requirements.txt
python-telegram-bot==20.8
pillow==10.4.0